+++ src/dftbp/dftbplus/initprogram.F90
@@ -10,11 +10,12 @@
 
 !> Global variables and initialization for the main program.
 module dftbp_dftbplus_initprogram
-  use dftbp_common_accuracy, only : dp, lc, mc, sc, elecTolMax, minTemp, tolSameDist, tolEfEquiv
+  use dftbp_common_accuracy, only : rsp, dp, lc, mc, sc, elecTolMax, minTemp, tolSameDist,&
+      & tolEfEquiv
   use dftbp_common_atomicmass, only : getAtomicMass
   use dftbp_common_coherence, only : checkToleranceCoherence, checkExactCoherence
   use dftbp_common_constants, only : shellNames, Hartree__eV, Bohr__AA, amu__au, pi, au__ps,&
-      & Bohr__nm, Hartree__kJ_mol, Boltzmann
+      & Bohr__nm, Hartree__kJ_mol, Boltzmann, AA__Bohr
   use dftbp_common_envcheck, only : checkStackSize
   use dftbp_common_environment, only : TEnvironment, globalTimers
   use dftbp_common_file, only : TFileDescr, setDefaultBinaryAccess, clearFile
@@ -78,6 +79,7 @@ module dftbp_dftbplus_initprogram
   use dftbp_elecsolvers_elsisolver, only : TElsiSolver_init, TElsiSolver_final
   use dftbp_extlibs_arpack, only : withArpack
   use dftbp_extlibs_elsiiface, only : withELSI
+  use dftbp_extlibs_externalmodel, only : TExtModel, externalModel_init
   use dftbp_extlibs_plumed, only : withPlumed, TPlumedCalc, TPlumedCalc_init
   use dftbp_extlibs_poisson, only : TPoissonInput
   use dftbp_extlibs_sdftd3, only : TSDFTD3, TSDFTD3_init, writeSDFTD3Info
@@ -118,6 +120,7 @@ module dftbp_dftbplus_initprogram
       & TMixerCmplx_init
   use dftbp_mixer_simplemixer, only : TSimpleMixerReal, TSimpleMixerCmplx, TSimpleMixerReal_init,&
       & TSimpleMixerCmplx_init
+  use dftbp_modelindependent_localclusters, only : TLocalClusters, TLocalClusters_init
   use dftbp_reks_reks, only : TReksInp, TReksCalc, reksTypes, REKS_init
   use dftbp_solvation_cm5, only : TChargeModel5, TChargeModel5_init
   use dftbp_solvation_fieldscaling, only : TScaleExtEField, init_TScaleExtEField
@@ -325,11 +328,20 @@ module dftbp_dftbplus_initprogram
     !> Hamiltonian type
     integer :: hamiltonianType
 
+    !> External model
+    type(TExtModel), allocatable :: extModel
+
+    !> Local environments around atomic sites
+    type(TLocalClusters), allocatable :: localClusters
+
     !> Raw H^0 hamiltonian data
     type(TSlakoCont) :: skHamCont
 
     !> Raw overlap hamiltonian data
     type(TSlakoCont) :: skOverCont
+
+    !> Is this an orthogonal basis used
+    logical :: isNonOrthogonal
 
     !> Repulsive (force-field like) interactions
     class(TRepulsive), allocatable :: repulsive
@@ -1345,7 +1357,7 @@ contains
 
     logical :: tGeoOptRequiresEgy, isOnsiteCorrected
     type(TStatus) :: errStatus
-    integer :: nLocalRows, nLocalCols
+    integer :: nLocalRows, nLocalCols, nAts
 
     logical :: isIoProc
 
@@ -1434,12 +1446,29 @@ contains
       end do
     end do
 
+
+  #:if WITH_SOCKETS
+    this%tSocket = allocated(input%ctrl%socketInput)
+    if (this%tSocket) then
+      this%tForces = .true.
+    end if
+  #:endif
+    this%tPrintForces = input%ctrl%tPrintForces
+    this%tForces = input%ctrl%tForces .or. this%tPrintForces
+
+    this%isNonOrthogonal = .true.
+
     select case(this%hamiltonianType)
     case default
-      call error("Invalid Hamiltonian")
+
+      call error("Invalid Model choice")
+
     case(hamiltonianTypes%dftb)
+
       this%orb = input%slako%orb
+
     case(hamiltonianTypes%xtb)
+
       if (.not.allocated(input%ctrl%tbliteInp)) then
         call error("xTB calculation supported only with tblite library")
       end if
@@ -1460,7 +1489,85 @@ contains
       ! resulting in a strange run-time error message. Turning it into move_alloc seems to avoid it.
       !this%orb = input%slako%orb
       call move_alloc(input%slako%orb, this%orb)
+
+    case(hamiltonianTypes%api)
+
+      if (.not.(input%ctrl%extModel%gives_dc_energy .and. input%ctrl%extModel%gives_ham)) then
+        if (this%tForces) then
+          call error("Force expressions are not available for this calculation")
+        end if
+      end if
+
+      call externalModel_init(this%extModel, input%ctrl%extModel, this%speciesName, errStatus)
+      if (errStatus%hasError()) then
+        call error(errStatus%message)
+      end if
+
+      if (input%ctrl%extModel%gives_ham) then
+
+        ! distance over which external model predicts matrix elements
+        this%cutOff%skCutOff = this%extModel%interactionCutoff
+        ! longer distance for producing a 'rounded cyclinder' containing nearby atoms, such that the
+        ! distance to a diatomic bond axis is at least this%cutOff%skCutOff
+        this%cutOff%mCutOff = max(this%cutOff%skCutOff,&
+            & sqrt(this%extModel%environmentCutoff**2 + (this%cutOff%skCutoff**2)/4.0_dp)&
+            & + epsilon(0.0_rsp))
+
+        allocate(this%orb)
+        allocate(this%orb%nshell(input%geom%nSpecies), source=0)
+        this%orb%nshell = this%extModel%nShellsOnSpecies
+        allocate(this%orb%nOrbSpecies(input%geom%nSpecies), source=0)
+        do iSp = 1, input%geom%nSpecies
+          do iSh = 1, this%orb%nshell(iSp)
+            this%orb%nOrbSpecies(iSp) = this%orb%nOrbSpecies(iSp)&
+                & + 2 * this%extModel%shellLValues(iSh,iSp) + 1
+          end do
+        end do
+        allocate(this%orb%nOrbAtom(this%nAtom))
+        this%orb%nOrbAtom(:) = this%orb%nOrbSpecies(this%species0)
+        allocate(this%orb%angShell(maxval(this%orb%nShell), input%geom%nSpecies))
+        this%orb%angShell(:,:) = 0
+        do iSp = 1, input%geom%nSpecies
+          do iSh = 1, this%orb%nshell(iSp)
+            this%orb%angShell(iSh,iSp) = this%extModel%shellLValues(iSh,iSp)
+          end do
+        end do
+
+        allocate(this%orb%iShellOrb(maxval(this%orb%nOrbSpecies), input%geom%nSpecies))
+        this%orb%iShellOrb(:,:) = 0
+        allocate(this%orb%posShell(maxval(this%orb%nshell) + 1, input%geom%nSpecies))
+        this%orb%posShell(:,:) = 0
+        do iSp = 1, input%geom%nSpecies
+          ii = 1
+          do iSh = 1, this%orb%nshell(iSp)
+            this%orb%posShell(iSh, iSp) = ii
+            this%orb%iShellOrb(ii:ii+2*this%extModel%shellLValues(iSh,iSp), iSp) = iSh
+            ii = ii + 2*this%extModel%shellLValues(iSh,iSp) + 1
+          end do
+          this%orb%posShell(this%orb%nshell(iSp) + 1, iSp) = ii
+        end do
+
+        this%orb%mShell = maxval(this%orb%nshell)
+        this%orb%mOrb = maxval(this%orb%nOrbSpecies)
+        this%orb%nOrb = sum(this%orb%nOrbAtom)
+
+        allocate(input%slako%skOcc(this%orb%mShell, input%geom%nSpecies))
+        input%slako%skOcc(:,:) = this%extModel%shellOccs
+
+        this%isNonOrthogonal = input%ctrl%extModel%gives_overlap
+
+      elseif (input%ctrl%extModel%gives_dc_energy) then
+
+        this%cutOff%mCutOff = this%extModel%interactionCutoff
+
+      else
+
+        call error("Model provides neither energy nor hamiltonian components")
+
+      end if
+
     end select
+
     this%nOrb = this%orb%nOrb
 
     ! start by assuming stress can be calculated if periodic
@@ -1613,7 +1720,7 @@ contains
       @:ASSERT(size(input%slako%mass) == this%nType)
       allocate(this%speciesMass(this%nType))
       this%speciesMass(:) = input%slako%mass(:)
-    case(hamiltonianTypes%xtb)
+    case(hamiltonianTypes%xtb, hamiltonianTypes%api)
       allocate(this%speciesMass(this%nType))
       this%speciesMass(:) = getAtomicMass(this%speciesName)
     end select
@@ -1662,6 +1769,8 @@ contains
     case(hamiltonianTypes%xtb)
       this%cutOff%skCutoff = this%tblite%getRCutoff()
       this%cutOff%mCutoff = this%cutOff%skCutoff
+    case(hamiltonianTypes%api)
+      ! Already set above
     end select
 
     call initHubbardUs_(input, this%orb, this%hamiltonianType, hubbU)
@@ -1947,11 +2056,9 @@ contains
     this%BarostatStrength = input%ctrl%BarostatStrength
 
   #:if WITH_SOCKETS
-    this%tSocket = allocated(input%ctrl%socketInput)
     if (this%tSocket) then
       input%ctrl%socketInput%nAtom = this%nAtom
       call this%initSocket(env, input%ctrl%socketInput)
-      this%tForces = .true.
       this%isGeoOpt = .false.
       this%tMD = .false.
     end if
@@ -1993,8 +2100,6 @@ contains
     this%tPrintEigVecs = input%ctrl%tPrintEigVecs
     this%tPrintEigVecsTxt = input%ctrl%tPrintEigVecsTxt
 
-    this%tPrintForces = input%ctrl%tPrintForces
-    this%tForces = input%ctrl%tForces .or. this%tPrintForces
     if (this%isLinResp) then
       allocate(this%linearResponse)
     end if
@@ -3078,6 +3183,19 @@ contains
     do iAt = 1, size(this%iAtInCentralRegion)
       this%iAtInCentralRegion(iAt) = iAt
     end do
+
+    nAts = size(this%iAtInCentralRegion)
+  #:if WITH_TRANSPORT
+    if (this%transpar%ncont > 0) then
+      nAts = this%transpar%contacts(this%transpar%ncont)%idxrange(2)
+    end if
+  #:endif
+
+    if (this%hamiltonianType == hamiltonianTypes%api) then
+      allocate(this%localClusters)
+      call TLocalClusters_init(this%localClusters, nAts, this%extModel%environmentCutoff,&
+          & this%cutOff%skCutoff)
+    end if
 
     if (allocated(this%eField)) then
       if (allocated(input%ctrl%atomicExtPotential)) then
@@ -4666,6 +4784,8 @@ contains
     case(hamiltonianTypes%xtb)
       ! It's an xTB Hamiltonian masquerading as a DFTB one (for now)
       referenceN0(:,:) = input%slako%skOcc(1:orb%mShell, :)
+    case(hamiltonianTypes%api)
+      referenceN0(:,:) = input%slako%skOcc(1:orb%mShell, :)
     end select
 
   end subroutine initReferencePopulation_
@@ -4741,7 +4861,8 @@ contains
       @:ASSERT(size(input%slako%skHubbU, dim=2) == nSpecies)
       hubbU(:,:) = input%slako%skHubbU(1:orb%mShell, :)
     case(hamiltonianTypes%xtb)
-      ! handled elsewhere
+      ! TODO
+    case(hamiltonianTypes%api)
     end select
 
     if (allocated(input%ctrl%hubbU)) then