+++ CHANGELOG.rst
@@ -11,6 +11,11 @@ Unreleased
 Added
 -----
 
+- External models (electronic for the moment) from linked (static or
+  dynamic) library outside the code base.
+
+- Rename "Hamiltonian" to "Model" in input, as more descriptive
+
 - Generalization of mixers to also handle complex density matrices
 
 - General range-separated, long-range corrected CAM hybrid functionals for
@@ -45,6 +50,12 @@ Fixed
   is not formally derived in the general case.
 
 - DeltaDFTB purified forces used correctly.
+
+Changed
+-------
+
+- Updated parser version to 15.
+
 
 24.1 (2024-02-12)
 =================